requirements.txt
# used
gcsfs==0.6.2
google-cloud-bigquery==1.24.0
google-cloud-storage==1.28.1
google-auth==1.16.0
google-cloud-bigquery-storage==0.8.0
pandas==1.0.3
pandas-gbq==0.13.1
pytz==2020.1
requests_toolbelt==0.9.1
